requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.35.0
diffusers>=1.26.0
accelerate>=0.24.0
peft>=0.5.0
safetensors>=0.3.1
Pillow>=10.0.0
opencv-python>=4.8.1
numpy>=1.25.0
tqdm>=4.66.1
scipy>=1.11.1
matplotlib>=3.8.1
pandas>=2.1.1
scikit-learn>=1.3.2
sentencepiece>=0.1.99
tokenizers>=0.14.0
xformers>=0.0.23
datasets>=2.14.5
notebook>=7.0.0
ipython>=8.15.0
